factor structs into separate test unit
